// src/oncobridge-web/src/app/api/models/finding-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { FindingTargetResponse } from '../models/finding-target-response';
export interface FindingResponse {
  actual?: (string | null);
  category: string;
  checkId: string;
  citation: string;
  expected?: (string | null);
  message: string;
  severity: string;
  target: FindingTargetResponse;
}